Fix: fixed logic to verify if an user already had commented in one specific annnounce
